Move struct `generator` into header `kernel/generator.hpp`
