+++ script/setup-macos.sh
@@ -1,6 +1,6 @@
 #!/bin/sh -e
 
-required()
+essential()
 {
   echo gmp
 }
@@ -9,13 +9,6 @@ optional()
 {
   echo gcc
   echo shellcheck # GitHub Actions
-}
-
-documentation()
-{
-  echo bibtex2html # script/references.sh
-  echo doxygen
-  echo pandoc # script/references.sh
 }
 
 brew update
@@ -27,10 +20,9 @@ else
   for each in "$@"
   do
     case "$each" in
-      -a | --all           ) ( required && optional && documentation ) | xargs brew install ;;
-      -d | --documentation ) (                         documentation ) | xargs brew install ;;
-      -o | --optional      ) (             optional                  ) | xargs brew install ;;
-      -r | --required      ) ( required                              ) | xargs brew install ;;
+      -a | --all       ) ( essential && optional ) | xargs brew install ;;
+      -e | --essential ) ( essential             ) | xargs brew install ;;
+      -o | --optional  ) (              optional ) | xargs brew install ;;
     esac
   done
 fi
