+++ script/setup-ubuntu.sh
@@ -1,6 +1,6 @@
 #!/bin/sh -e
 
-required()
+essential()
 {
   echo build-essential
   echo libgmp-dev
@@ -8,19 +8,8 @@ required()
 
 optional()
 {
-  echo kcachegrind # script/callgrind.sh
-  echo massif-visualizer # script/massif.sh
   echo shellcheck # GitHub Actions
   echo valgrind
-}
-
-documentation()
-{
-  wget -q https://github.com/jgm/pandoc/releases/download/3.3/pandoc-3.3-1-amd64.deb -P /tmp
-
-  echo bibtex2html # script/references.sh
-  echo doxygen
-  echo /tmp/pandoc-3.3-1-amd64.deb # script/references.sh
 }
 
 sudo apt update
@@ -32,10 +21,9 @@ else
   for each in "$@"
   do
     case "$each" in
-      -a | --all           ) ( required && optional && documentation ) | xargs sudo apt install --yes ;;
-      -d | --documentation ) (                         documentation ) | xargs sudo apt install --yes ;;
-      -o | --optional      ) (             optional                  ) | xargs sudo apt install --yes ;;
-      -r | --required      ) ( required                              ) | xargs sudo apt install --yes ;;
+      -a | --all       ) ( essential && optional ) | xargs sudo apt install --yes ;;
+      -e | --essential ) ( essential             ) | xargs sudo apt install --yes ;;
+      -o | --optional  ) (              optional ) | xargs sudo apt install --yes ;;
     esac
   done
 fi
