+++ src/kernel/syntactic_environment.cpp
@@ -17,270 +17,13 @@
 #include <meevax/kernel/boolean.hpp>
 #include <meevax/kernel/environment.hpp>
 #include <meevax/kernel/expander.hpp>
+#include <meevax/kernel/generator.hpp>
 #include <meevax/kernel/identity.hpp>
 #include <meevax/kernel/syntactic_environment.hpp>
 #include <meevax/kernel/transformer.hpp>
 
 namespace meevax::inline kernel
 {
-  #define GENERATOR(NAME)                                                      \
-  auto NAME([[maybe_unused]] syntactic_environment & generator,                \
-            [[maybe_unused]] object const& form,                               \
-            [[maybe_unused]] object const& bound_variables,                    \
-            [[maybe_unused]] object const& continuation,                       \
-            [[maybe_unused]] bool tail) -> object
-
-  GENERATOR(syntactic_environment::generator::quote)
-  {
-    return cons(make(instruction::secd_load_constant), car(form).is<syntactic_closure>() ? car(form).as<syntactic_closure>().form
-                                                                                         : car(form),
-                continuation);
-  }
-
-  GENERATOR(syntactic_environment::generator::quote_syntax)
-  {
-    return cons(make(instruction::secd_load_constant), car(form),
-                continuation);
-  }
-
-  GENERATOR(syntactic_environment::generator::call)
-  {
-    return operand(generator,
-                   cdr(form),
-                   bound_variables,
-                   generator.generate(car(form),
-                                      bound_variables,
-                                      tail ? list(make(instruction::secd_tail_call))
-                                           : cons(make(instruction::secd_call), continuation)));
-  }
-
-  GENERATOR(syntactic_environment::generator::operand)
-  {
-    if (form.is<pair>())
-    {
-      return operand(generator,
-                     cdr(form),
-                     bound_variables,
-                     generator.generate(car(form),
-                                        bound_variables,
-                                        cons(make(instruction::secd_cons),
-                                             continuation)));
-    }
-    else
-    {
-      return generator.generate(form, bound_variables, continuation);
-    }
-  }
-
-  GENERATOR(syntactic_environment::generator::lambda)
-  {
-    return cons(make(instruction::secd_load_closure),
-                body(generator,
-                     cdr(form),
-                     cons(car(form), bound_variables), // Extend scope.
-                     list(make(instruction::secd_return))),
-                continuation);
-  }
-
-  GENERATOR(syntactic_environment::generator::body)
-  {
-    if (cdr(form).template is<null>())
-    {
-      return generator.generate(car(form),
-                                bound_variables,
-                                continuation,
-                                true);
-    }
-    else
-    {
-      return generator.generate(car(form),
-                                bound_variables,
-                                cons(make(instruction::secd_drop),
-                                     body(generator,
-                                          cdr(form),
-                                          bound_variables,
-                                          continuation)),
-                                false);
-    }
-  }
-
-  GENERATOR(syntactic_environment::generator::conditional)
-  {
-    if (tail)
-    {
-      assert(lexical_cast(continuation) == "(return)");
-
-      return generator.generate(car(form), // <test>
-                                bound_variables,
-                                list(make(instruction::secd_tail_select),
-                                     generator.generate(cadr(form),
-                                                        bound_variables,
-                                                        continuation,
-                                                        tail),
-                                     cddr(form) ? generator.generate(caddr(form),
-                                                                     bound_variables,
-                                                                     continuation,
-                                                                     tail)
-                                                : list(make(instruction::secd_load_constant), unspecified, // If <test> yields a false value and no <alternate> is specified, then the result of the expression is unspecified.
-                                                       make(instruction::secd_return))));
-    }
-    else
-    {
-      return generator.generate(car(form), // <test>
-                                bound_variables,
-                                cons(make(instruction::secd_select),
-                                     generator.generate(cadr(form),
-                                                        bound_variables,
-                                                        list(make(instruction::secd_join))),
-                                     cddr(form) ? generator.generate(caddr(form),
-                                                                     bound_variables,
-                                                                     list(make(instruction::secd_join)))
-                                                : list(make(instruction::secd_load_constant), unspecified, // If <test> yields a false value and no <alternate> is specified, then the result of the expression is unspecified.
-                                                       make(instruction::secd_join)),
-                                     continuation));
-    }
-  }
-
-  GENERATOR(syntactic_environment::generator::set)
-  {
-    assert(car(form).is_also<identifier>());
-
-    if (let const& identity = generator.identify(car(form), bound_variables); identity.is<relative>())
-    {
-      return generator.generate(cadr(form),
-                                bound_variables,
-                                cons(make(instruction::secd_store_relative), identity,
-                                     continuation));
-    }
-    else if (identity.is<variadic>())
-    {
-      return generator.generate(cadr(form),
-                                bound_variables,
-                                cons(make(instruction::secd_store_variadic), identity,
-                                     continuation));
-    }
-    else
-    {
-      assert(identity.is<absolute>());
-
-      return generator.generate(cadr(form),
-                                bound_variables,
-                                cons(make(instruction::secd_store_absolute), identity,
-                                     continuation));
-    }
-  }
-
-  GENERATOR(syntactic_environment::generator::letrec)
-  {
-    assert(not tail or lexical_cast(continuation) == "(return)");
-
-    let const formals = map(car, car(form));
-
-    return cons(make(instruction::secd_dummy),
-                operand(generator,
-                        map(cadr, car(form)),
-                        cons(formals, bound_variables),
-                        lambda(generator,
-                               cons(formals, cdr(form)), // (<formals> <body>)
-                               bound_variables,
-                               tail ? list(make(instruction::secd_tail_letrec))
-                                    : cons(make(instruction::secd_letrec), continuation))));
-  }
-
-  GENERATOR(syntactic_environment::generator::sequence)
-  {
-    if (cdr(form).is<null>()) // is tail sequence
-    {
-      return generator.generate(car(form),
-                                bound_variables,
-                                continuation,
-                                tail);
-    }
-    else
-    {
-      /*
-         The top-level sequential expression may contain macro definitions.
-         In that case, the macro definition must be compiled before the
-         macro is used (the evaluation order of function arguments in C++
-         is not specified, but in most environments they are evaluated from
-         right to left). Therefore, the first expression is compiled
-         separately and then combined with the compiled result of the
-         remaining expressions by append.
-      */
-      let const& head = generator.generate(car(form), // Head expression or definition
-                                           bound_variables,
-                                           unit);
-      return append(head,
-                    cons(make(instruction::secd_drop), // Pop result of head expression
-                         sequence(generator,
-                                  cdr(form), // Rest expression or definitions
-                                  bound_variables,
-                                  continuation,
-                                  tail)));
-    }
-  }
-
-  GENERATOR(syntactic_environment::generator::define)
-  {
-    assert(not car(form).is<pair>()); // This has been checked on previous passes.
-
-    assert(car(form).is_also<identifier>());
-
-    if (bound_variables)
-    {
-      throw error(make<string>("definition cannot appear in this syntactic-context"));
-    }
-    else
-    {
-      return generator.generate(cdr(form) ? cadr(form) : unspecified,
-                                bound_variables,
-                                cons(make(instruction::secd_store_absolute), generator.identify(car(form), bound_variables),
-                                     continuation));
-    }
-  }
-
-  GENERATOR(syntactic_environment::generator::define_syntax)
-  {
-    assert(car(form).is_also<identifier>());
-
-    let identity = generator.identify(car(form), unit);
-
-    cdr(identity) = make<transformer>(environment().execute(generator.generate(cadr(form),
-                                                                               bound_variables)),
-                                      make<syntactic_environment>(bound_variables,
-                                                                  generator.second));
-
-    return cons(make(instruction::secd_load_constant), unspecified,
-                continuation);
-  }
-
-  GENERATOR(syntactic_environment::generator::call_with_current_continuation)
-  {
-    assert(form.is<pair>());
-    assert(cdr(form).is<null>());
-
-    return cons(make(instruction::secd_load_continuation),
-                continuation,
-                generator.generate(car(form),
-                                   bound_variables,
-                                   list(make(instruction::secd_tail_call)), // The first argument passed to call-with-current-continuation must be called via a tail call.
-                                   tail));
-  }
-
-  GENERATOR(syntactic_environment::generator::current)
-  {
-    return cons(make(instruction::secd_current), car(form),
-                continuation);
-  }
-
-  GENERATOR(syntactic_environment::generator::install)
-  {
-    return generator.generate(cadr(form),
-                              bound_variables,
-                              cons(make(instruction::secd_install), car(form),
-                                   continuation));
-  }
-
   auto syntactic_environment::compile(object const& form) -> object
   {
     return generate(expand(form, first), first);
